moved MapFeatures to seperate file
